requirements.txt
boto3>=1.34.0
requests>=2.31.0
aws-requests-auth>=0.4.3
python-dateutil>=2.8.2
PyYAML